Reworked mutex handling + updated C++ template for a var init
- Mutexes have been consolidated. Instead of one per DFA (which can easily get to hundreds of them) we only have one mutex in the Recognizer class and all other parties use this for serialization. It's only about protected the DFA anyway, which is stored in a recognizer (lexer/parser).
- ATNState::getStateType() returns a size_t value now (actually an enum).
- Replaced checks via RTTI for transitions by the (serialization) type of the transition, for simplicity.
- Added some missing initialization for fields in certain ATN state classes.
- Fixed mem leak in DFA by shadowing the s0 field. That way still have a ref to the self created instance, even is s0 was replaced later.
- Added variable init in code generation for a rule context declaration (e.g. for labels).

+++ runtime/Cpp/runtime/src/atn/ATNState.h
@@ -123,7 +123,6 @@ namespace atn {
 
     static const std::vector<std::string> serializationNames;
 
-    /// Which ATN are we in?
     size_t stateNumber = INVALID_STATE_NUMBER;
     size_t ruleIndex = 0; // at runtime, we don't have Rule objects
     bool epsilonOnlyTransitions = false;
@@ -143,7 +142,7 @@ namespace atn {
     virtual void addTransition(Transition *e);
     virtual void addTransition(size_t index, Transition *e);
     virtual Transition* removeTransition(size_t index);
-    virtual int getStateType() = 0;
+    virtual size_t getStateType() = 0;
   };
 
 } // namespace atn
