+++ runtime/Cpp/runtime/src/atn/BlockEndState.cpp
@@ -36,6 +36,6 @@ using namespace antlr4::atn;
 BlockEndState::BlockEndState() : startState(nullptr) {
 }
 
-int BlockEndState::getStateType() {
+size_t BlockEndState::getStateType() {
   return BLOCK_END;
 }
