+++ runtime/Cpp/runtime/src/atn/RuleStopState.h
@@ -36,16 +36,14 @@
 namespace antlr4 {
 namespace atn {
 
-  /// <summary>
   /// The last node in the ATN for a rule, unless that rule is the start symbol.
-  ///  In that case, there is one transition to EOF. Later, we might encode
-  ///  references to all calls to this rule to compute FOLLOW sets for
-  ///  error handling.
-  /// </summary>
+  /// In that case, there is one transition to EOF. Later, we might encode
+  /// references to all calls to this rule to compute FOLLOW sets for
+  /// error handling.
   class ANTLR4CPP_PUBLIC RuleStopState final : public ATNState {
 
   public:
-    virtual int getStateType() override;
+    virtual size_t getStateType() override;
 
   };
